cdk/python/glue/requirements.txt
aws-cdk-lib==2.96.2
constructs>=10.0.0,<11.0.0
